updated extraction of Q to Ratio As a result, saw (Show instance of Ratio) the complete output of the robot program when extracted to Haskell:
*RoboExtract> RoboExtract.robotOutputShowable
[(0 % 1,(0 % 1,1 % 2)),(1571 % 1000,(0 % 1,0 % 1)),(1 % 1,(1 % 10,0 % 1)),(7071 % 500,(0 % 1,0 % 1))]

This semantically matched with the output in Coq:

Eval vm_compute in robotOutputShowable.
     = [(0, (0, 1 # 2)); (1571 # 1000, (0, 0)); (1, (1 # 10, 0));
       (14142 # 1000, (0, 0))]
     : list (Q and Q and Q)

There is a syntactical difference. The Haskell version is normalized, e.g. 7071 % 500 instead of 14142 # 1000
I was not expecting this difference.

Also, I had to manually add "import Data.Ratio" in the extract and fix the previously known compilation error.