started defining primitive ROS types in Coq.
It seems that it is impossible to define the stream merging operator in Coq.
Such an operator will have to predict on which stream the next message will arrive.
This is the stream on which we pattern match first.
If the guess is wrong, the message at the other stream might be unnecessarily delayed.

Also, to be a function, the result of merging 2 streams should not depend on timing factors.
Perhaps the Monad needs to be moved inside the stream defn, just as the roshask definition of Topic.
Any issues with strict positivity?