fixed the root cause of rewrite failures, as identified in https://github.com/aa755/ROSCoq/commit/05970005b82822e7b867cb2f9aba4fbed0fc2307
While reading the description of "Existing Instance" in the coq manual,
where they say that it is just a sugar Hint Resolve,
realized that objects that are not typeclass instances can also be added to the Hint DB used
for typeclass resolution (typeclass_instances).

Thus, when reasoning about a particular CPS execution, the user just has to
add that executuion to typeclass_instances.
Then, Coq will know what event ordering the user is talking about.