fixed rewriting, suff. condition for the car to be confined during a sequence of atomic moves.
lesson learnt:
for a function f of type A->B->C, with
non trivial equality  on either of A, B, C,
there should be an instance of the form
Proper (equiv ==> equiv ==> equiv) f
Each of A and B and C should have an Equiv instance.

In many cases, I was lazy, and did not define an
Equiv instance on a middle argument (e.g. B),
and proved a lemma like

forall b,
Proper (equiv ==> equiv) (fun a  => f a b)