the optimizer typechecks, is hopefully transparent enough now to run.
Also, got rid of the dependent argument needed to prove positivity of cos,
in order to take its reciprocal. In the relevant part of
the solution space,
this cos term was bounded below by a positive value. So
took the max with that positive value.

Up next: subdivide the solution space into equi-spaced (rational?)
points, and
feed the list of points to the optimizer, which will pick the best of
safe ones among those points.

If it runs, start working on proving that if the subdivision is fine enough,
the optimizer is guaranteed to return not only an answer, but an approximately
optimal one.