started error analysis of position for perfect straight case.
the previous analysis assumes turnCurvature <> 0, so that division by turnCurvature
is allowed -- it shows up even in the statement, not just in the proof.
Therefore, the turnCurvature=0(absolute straight motion) case has to be handled separately.

Because were allowing some error in turn curvature, can the turnCurvature=0 be handled
as turnCurvature=eps for some very small epsilon?